pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pytorch-cpr"
version = "0.3.0"
authors = [
  { name="Jörg Franke", email="frankej@cs.uni-freiburg.de" },
]
description = "Constrained Parameter Regularization for PyTorch"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
full = [
    'pytorch>=2.3.1',
]

[project.urls]
Homepage = "https://github.com/automl/CPR"


# update pip package
# https://packaging.python.org/en/latest/guides/distributing-packages-using-setuptools/
#> pip install build twine
#> python3 -m build --sdist
#> python3 -m build --wheel
#> twine check dist/*
#> twine upload dist/*